Remove BOOTSEL check in debug mode

+++ fw-next/ticker.h
@@ -12,7 +12,7 @@ public:
     void reset(uint32_t tick_duration);
 
 private:
-    uint8_t tick_duration;
+    uint32_t tick_duration;
     uint32_t last_tick_time = 0;
     bool first_tick = true;
 };
